pyproject.toml
[project]
name = "wf-firmware-uploader"
version = "0.1.0"
description = "GUI firmware uploader for STM32 and ESP32 microcontrollers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.5.0",
    "pyserial>=3.5",
    "esptool>=4.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pylint>=2.17.0",
    "mypy>=1.5.0",
    "types-pyserial>=3.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# Flake8 configuration
[tool.flake8]
max-line-length = 100
ignore = ["E501", "W503", "E203", "F401"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    ".eggs",
    "*.egg",
    "build",
    "dist"
]
per-file-ignores = [
    "__init__.py:F401",
]

# Pylint configuration
[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    "R0903",  # too-few-public-methods
    "R0913",  # too-many-arguments
    "R0917",  # too-many-positional-arguments
    "E0611",  # no-name-in-module (for PySide6 dynamic imports)
    "E0401",  # import-error
]

[tool.pylint.format]
max-line-length = 100

[tool.pylint]
extension-pkg-whitelist = ["PySide6"]

# MyPy configuration
[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
show_error_codes = true
no_strict_optional = true
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
